+++ source/global.d.ts
@@ -108,7 +108,6 @@ declare module global {
   var translations: any
   var targets: TargetProvider
   var tags: TagProvider
-  var links: LinkProvider
   var stats: StatsProvider
   var recentDocs: RecentDocumentsProvider
   // Translation data necessary to facilitate internationalisation