+++ source/main/modules/fsal/fsal-file.ts
@@ -29,6 +29,7 @@ import extractBOM from './util/extract-bom'
 import extractTags from './util/extract-tags'
 import extractLinks from './util/extract-links'
 import { SearchTerm } from '@dts/common/search'
+import { removeLinks, reportLinks } from '@providers/link-provider'
 
 // Here are all supported variables for Pandoc:
 // https://pandoc.org/MANUAL.html#variables
@@ -277,7 +278,7 @@ export async function parse (filePath: string, cache: FSALCache|null, parent: Di
 
   // Finally, report the tags
   global.tags.report(file.tags, file.path)
-  global.links.report(file.path, file.links, file.id)
+  reportLinks(file.path, file.links, file.id)
 
   return file
 }
@@ -348,10 +349,10 @@ export async function save (fileObject: MDFileDescriptor, content: string, cache
   await updateFileMetadata(fileObject)
   // Make sure to keep the file object itself as well as the tags updated
   global.tags.remove(fileObject.tags, fileObject.path)
-  global.links.remove(fileObject.path, fileObject.id)
+  removeLinks(fileObject.path, fileObject.id)
   parseFileContents(fileObject, content)
   global.tags.report(fileObject.tags, fileObject.path)
-  global.links.report(fileObject.path, fileObject.links, fileObject.id)
+  reportLinks(fileObject.path, fileObject.links, fileObject.id)
   fileObject.modified = false // Always reset the modification flag.
   if (cache !== null) {
     cacheFile(fileObject, cache)
@@ -429,10 +430,10 @@ export async function reparseChangedFile (fileObject: MDFileDescriptor, cache: F
   await updateFileMetadata(fileObject)
   // Make sure to keep the file object itself as well as the tags updated
   global.tags.remove(fileObject.tags, fileObject.path)
-  global.links.remove(fileObject.path, fileObject.id)
+  removeLinks(fileObject.path, fileObject.id)
   parseFileContents(fileObject, contents)
   global.tags.report(fileObject.tags, fileObject.path)
-  global.links.report(fileObject.path, fileObject.links, fileObject.id)
+  reportLinks(fileObject.path, fileObject.links, fileObject.id)
   fileObject.modified = false // Always reset the modification flag.
   if (cache !== null) {
     cacheFile(fileObject, cache)