+++ CHANGELOG.md
@@ -52,6 +52,7 @@
 - Polyfill the `path` module in renderer processes
 - Transform the CSS provider into a functional module
 - Transform the appearance provider into a functional module
+- Transform the link provider into a functional module
 - Convert the MainSidebar component to TypeScript
 - Builds are now done without type checking, increasing performance